kio: new KFilePreviewJob class
including ported a port of KImageFilePreview to it. question is do I
want to support directory and remote file(s) previews, do I want it to
emit QPixmap when the preview plugins (including the current KIO
thumbnailers) create QImage

+++ kio/kio/kfilepreview.cpp
@@ -47,7 +47,7 @@ static QStringList kPreviewGlobMimeTypes(const QStringList &servicetypes)
     return result;
 }
 
-static void kPreviewOverlay(QImage &preview, const QImage &overlay, const ushort overlayalpha)
+static void kPreviewOverlay(QImage &preview, const QImage &overlay, const int overlayalpha)
 {
     // TODO: ratio check
     QPainter painter(&preview);
